pyproject.toml
[tool.poetry]
name = "floes"
version = "0.0.1"
description = ""
authors = ["dbisk <16860792+dbisk@users.noreply.github.com>"]
packages = [
    { include = "floe", from = "src/" }
]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.2"
torch = "^1.10.2"
grpcio = "^1.44.0"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.44.0"
torchvision = "^0.11.3"
tqdm = "^4.62.3"
torchaudio = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
